New feature: Rugs!
Development of the rugs are now done. The user is now able to drag and place a maximum of four rugs (more if you edit code). The user also can move rugs around on the screen and drag the rugs back to the starting area to discard it. Added a placeholder rug image.