Main menu update
Updated the main menu so that instead of showing plain text, it shows images for the title, and the "play" and "options" buttons. Also changed the code so that the user needs to click "play" instead of clicking anywhere. Also added a fading effect to the buttons - when the mouse is hovering over these buttons, an yellow outline will fade in. Move the mouse out, and the outline will fade out.